wind fix

+++ examples/flocking/behaviours/WindRule.cpp
@@ -10,7 +10,7 @@ Vector2 WindRule::computeForce(const std::vector<Boid*>& neighborhood, Boid* boi
     //return Vector2 of the wind angle (use trig rules)
     //Vector2(cos(x),sin(x))
 
-    return Vector2(cos(DEG2RAD(windAngle)), sin(DEG2RAD(windAngle)));
+    return Vector2((float)cos(DEG2RAD(windAngle)), (float)sin(DEG2RAD(windAngle)));
 
     //return Vector2::zero();
 }
